Use similarty framework to determine if two tracks are (likely) the same
Names of titles and artists come from different sources, E.g. various
music charts or the entries in a spotify catalog, each with their own
peculiarities in spelling etc. To check whether a track already is
present in a playlist, exact matches will lead to too many false
negatives, resulting easily in a large number of duplicates being added
to a playlist on spotify. Therefore, when comparing artist and title, we
must look at similarity match and not exact match.

I make use of the info.debatty:java-string-similarity:1.1.0 library.