Convert back from Kotlin to Java 17
Made the decision not to spent any more time on learning
Kotlin. I don't use it often enough and I forget all the
language details when I continue on this project, so it's not
productive for me. Also, I think it has not many features to
offer for this project compared to a Java 17 implementation.