Put server-side Spotify code behind a spring profile
The code requires additional setup, like a spotify.properties file. By
putting them behind a spring Profile, this code is not loaded on startup
but I also do not have to through it away yet.

To be complete: the spotify interaction has been moved from backend to
frontend. Therefore, it is no longer needed server-side.